compiler fix (SWITCHED TO JDK 11); moved everything to test -> java;
changed scenario name
plugin ALMOST added
