+++ .gitignore
@@ -1,1 +1,3 @@
-.idea
+.idea/
+bin/
+seleniumIPR.iml
