+++ src/test/java/Config/Utilities/AllureListener.java
@@ -1,4 +1,4 @@
-package HelpfulStuff;
+package Config.Utilities;
 
 import io.qameta.allure.listener.StepLifecycleListener;
