+++ src/test/java/Run/CucumberRunner.java
@@ -1,24 +1,23 @@
-package PageObject;
+package Run;
 
 import cucumber.api.CucumberOptions;
 import cucumber.api.testng.AbstractTestNGCucumberTests;
 import org.testng.annotations.AfterSuite;
 import org.testng.annotations.BeforeSuite;
 
-import static Config.Utilities.WebDriver.webDriver;
-
 //TODO: move browser config to before
 // and set up window handler
 
 @CucumberOptions(
         features = {"src/test/resources/features"},
-        glue = {"PageObject"},
-        tags = "@Test")
+        glue = {"Run"},
+//        plugin = {"io.qameta.allure.cucumber4jvm.AllureCucumber4Jvm"},
+        tags = "@Gmail")
 public class CucumberRunner extends AbstractTestNGCucumberTests {
+
 
     @BeforeSuite
     public void setUp() {
-
     }
 
 
@@ -26,4 +25,6 @@ public class CucumberRunner extends AbstractTestNGCucumberTests {
     public void tearDown() {
 //        webDriver.close();
     }
+
+
 }
