+++ src/test/resources/features/SendLetterGmail.feature
@@ -1,6 +1,6 @@
 Feature: MainPage
 
-  @Test
+  @Gmail
   Scenario: 01 - Main page operations
 
     Given test host open
