+++ testng.xml
@@ -1,0 +1,9 @@
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+<suite name="seleniumIPR" verbose="1">
+    <test name="Gmail">
+        <classes>
+            <class name="Run.CucumberRunner">
+            </class>
+        </classes>
+    </test>
+</suite>
